03-tools-and-setup/technology-stack/01-python-projects/pyproject.toml
[project]
name = "mini-etl-env"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pathlib (>=1.0.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
